+++ backend_restaurantmanager/src/main/java/com/tfg_rm/backend_restaurantmanager/auth/repository/AuthEmployeeRepository.java
@@ -11,7 +11,7 @@ import com.tfg_rm.backend_restaurantmanager.shared.entity.EmployeeEntity;
  * Repository for accessing employee records.
  */
 @Repository
-public interface AuthEmployeeRepository extends JpaRepository<EmployeeEntity, Integer> {
+public interface AuthEmployeeRepository extends JpaRepository<EmployeeEntity, Long> {
 
     Optional<EmployeeEntity> findByCode(String code);
