+++ backend_restaurantmanager/src/main/java/com/tfg_rm/backend_restaurantmanager/employee/repository/RestaurantTableOrderViewRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import com.tfg_rm.backend_restaurantmanager.shared.entity.view.RestaurantTableOrderViewEntity;
 
 @Repository
-public interface RestaurantTableOrderViewRepository extends JpaRepository<RestaurantTableOrderViewEntity, Integer> {
-    List<RestaurantTableOrderViewEntity> findByRestaurantId(Integer restaurantId);
+public interface RestaurantTableOrderViewRepository extends JpaRepository<RestaurantTableOrderViewEntity, Long> {
+    List<RestaurantTableOrderViewEntity> findByRestaurantId(Long restaurantId);
 }
